+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Big things planned! More unit tests! Less ORM quirks! More awesomeness!
 * **server:** print Miter's package version when starting a server
 * **jwt:** allow the JWT policies to specify where to find the authorization token
 * **crud-controller:** allow custom implementations of performDestroy
+* **router:** add transformRoute hook to allow controllers to programmatically modify or skip routes
 
 ### Bug Fixes
 
@@ -26,6 +27,7 @@ Big things planned! More unit tests! Less ORM quirks! More awesomeness!
 
 * **jwt:** allow the JWT policies to specify where to find the authorization token
 * **crud-controller:** allow custom implementations of performDestroy
+* **router:** add transformRoute hook to allow controllers to programmatically modify or skip routes
 
 ### Bug Fixes